[Task-893]-Change font size and color.Change header background color.
